Refactor: infoForm을 InfoCard로 변경
- ModifyInfoCard와 통일성을 위해 InfoCard로 변경했습니다.
