+++ client/src/components/profile/InfoCard/style.ts
@@ -1,0 +1,45 @@
+import styled, { css } from 'styled-components';
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 45rem;
+`;
+
+export const InfoWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin: 3px;
+`;
+export const Legend = styled.legend`
+  font-size: 3rem;
+  color: ${({ theme }) => theme.colors.bgGreen};
+`;
+
+export const Info = styled.div`
+  font-size: 5rem;
+  color: ${({ theme }) => theme.colors.bgGreen};
+  width: 40rem;
+  word-break: break-all;
+`;
+
+export const InfoSet = styled.fieldset`
+  ${({ theme }) => css`
+    background-color: ${theme.colors.grey};
+    padding: ${theme.paddings.lg};
+    border: 1px solid ${theme.colors.borderGrey};
+    border-radius: ${theme.borderRadius.base};
+  `};
+`;
+
+export const ModalToggleBtn = styled.span`
+  ${({ theme }) => theme.flexCenter}
+  background-color: ${({ theme }) => theme.colors.green};
+  padding: ${({ theme }) => theme.paddings.sm};
+  color: ${({ theme }) => theme.colors.white};
+  border-radius: 1rem;
+  cursor: pointer;
+  :hover {
+    color: ${({ theme }) => theme.colors.primary};
+  }
+`;
